requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pydantic==2.9.2
PyMySQL==1.1.1
pypdf==5.1.0
openai==1.51.2
pinecone-client==5.0.1
httpx==0.27.2
python-multipart==0.0.12
